Move RuleProps and QueryObject interfaces to a separate file
